+++ packages/ai/core/generate-text/generate-text-result.ts
@@ -3,11 +3,11 @@ import { CoreTool } from '../tool/tool';
 import {
   CallWarning,
   FinishReason,
-  LanguageModelRequestMetadata,
-  LanguageModelResponseMetadataWithHeaders,
   LogProbs,
   ProviderMetadata,
 } from '../types';
+import { LanguageModelRequestMetadata } from '../types/language-model-request-metadata';
+import { LanguageModelResponseMetadata } from '../types/language-model-response-metadata';
 import { LanguageModelUsage } from '../types/usage';
 import { StepResult } from './step-result';
 import { ToolCallArray } from './tool-call';
@@ -49,12 +49,7 @@ export interface GenerateTextResult<TOOLS extends Record<string, CoreTool>> {
   readonly warnings: CallWarning[] | undefined;
 
   /**
-The response messages that were generated during the call. It consists of an assistant message,
-potentially containing tool calls.
-
-When there are tool results, there is an additional tool message with the tool results that are available.
-If there are tools that do not have execute functions, they are not included in the tool results and
-need to be added separately.
+@deprecated use `response.messages` instead.
      */
   readonly responseMessages: Array<CoreAssistantMessage | CoreToolMessage>;
 
@@ -93,7 +88,17 @@ Additional request information.
   /**
 Additional response information.
    */
-  readonly response: LanguageModelResponseMetadataWithHeaders;
+  readonly response: LanguageModelResponseMetadata & {
+    /**
+The response messages that were generated during the call. It consists of an assistant message,
+potentially containing tool calls.
+
+When there are tool results, there is an additional tool message with the tool results that are available.
+If there are tools that do not have execute functions, they are not included in the tool results and
+need to be added separately.
+       */
+    messages: Array<CoreAssistantMessage | CoreToolMessage>;
+  };
 
   /**
 Logprobs for the completion.
