+++ packages/ai/core/generate-text/step-result.ts
@@ -3,7 +3,7 @@ import {
   CallWarning,
   FinishReason,
   LanguageModelRequestMetadata,
-  LanguageModelResponseMetadataWithHeaders,
+  LanguageModelResponseMetadata,
   LogProbs,
   ProviderMetadata,
 } from '../types';
@@ -71,7 +71,7 @@ Additional request information.
   /**
 Additional response information.
 */
-  readonly response: LanguageModelResponseMetadataWithHeaders;
+  readonly response: LanguageModelResponseMetadata;
 
   /**
 Additional provider-specific metadata. They are passed through
