+++ packages/ai/core/generate-text/stream-text.ts
@@ -1,7 +1,3 @@
-import {
-  LanguageModelV1Message,
-  LanguageModelV1Prompt,
-} from '@ai-sdk/provider';
 import { createIdGenerator } from '@ai-sdk/provider-utils';
 import { Span } from '@opentelemetry/api';
 import { ServerResponse } from 'node:http';
@@ -17,10 +13,7 @@ import {
 import { createResolvablePromise } from '../../util/create-resolvable-promise';
 import { retryWithExponentialBackoff } from '../../util/retry-with-exponential-backoff';
 import { CallSettings } from '../prompt/call-settings';
-import {
-  convertToLanguageModelMessage,
-  convertToLanguageModelPrompt,
-} from '../prompt/convert-to-language-model-prompt';
+import { convertToLanguageModelPrompt } from '../prompt/convert-to-language-model-prompt';
 import { prepareCallSettings } from '../prompt/prepare-call-settings';
 import { prepareToolsAndToolChoice } from '../prompt/prepare-tools-and-tool-choice';
 import { Prompt } from '../prompt/prompt';
@@ -869,7 +862,9 @@ class DefaultStreamTextResult<TOOLS extends Record<string, CoreTool>>
                 usage: stepUsage,
                 experimental_providerMetadata: stepProviderMetadata,
                 logprobs: stepLogProbs,
-                response: stepResponse,
+                response: {
+                  ...stepResponse,
+                },
                 isContinued: nextStepType === 'continue',
               });
 
@@ -975,7 +970,9 @@ class DefaultStreamTextResult<TOOLS extends Record<string, CoreTool>>
                   usage: combinedUsage,
                   experimental_providerMetadata: stepProviderMetadata,
                   logprobs: stepLogProbs,
-                  response: stepResponse,
+                  response: {
+                    ...stepResponse,
+                  },
                 });
 
                 // close the stitchable stream
