+++ packages/ai/core/types/language-model-request-metadata.ts
@@ -1,0 +1,6 @@
+export type LanguageModelRequestMetadata = {
+  /**
+  Raw request HTTP body that was sent to the provider API as a string (JSON should be stringified).
+     */
+  body?: string;
+};
